Move IPIRI and UMTSIRI job generation code out of collector_sync.c
This should make the collector_sync code a bit less cluttered.

Also tidied up the job generation functions, separating jobs
derived from packet events and jobs derived from external events
(e.g. changes to intercept configuration).

Fixed bug where some session-derived IPIRIs were missing an
appropriate event type.
